fixed bugs but there are new ones
