+++ src/features/markets/CryptoRow.tsx
@@ -35,6 +35,8 @@ export default function CryptoRow({
   const [anchorEl, setAnchorEl] = useState<null | HTMLElement>(null);
   const open = Boolean(anchorEl);
   const handleClick = (event: React.MouseEvent<HTMLButtonElement>) => {
+    event.stopPropagation();
+    event.preventDefault();
     setAnchorEl(event.currentTarget);
   };
   const handleClose = () => {
