+++ src/pages/Markets.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { useAllCryptoPrice } from "../features/markets/useAllCryptoPrice";
 import TopCryptoCard from "../features/markets/TopHotCryptoCard";
 import Spinner from "../ui/Spinner";
@@ -17,9 +17,9 @@ export default function Markets() {
         <>
           <div className="flex flex-col">
             <div className="grid grid-cols-1 lg:grid-cols-3 gap-4 px-8 w-full ">
-              <TopCryptoCard type="hot24" />
-              <TopCryptoCard type="top24" />
-              <TopCryptoCard type="big24" />
+              <TopCryptoCard data={data} type="hot24" />
+              <TopCryptoCard data={data} type="top24" />
+              <TopCryptoCard data={data} type="big24" />
             </div>
           </div>
           <div className="flex justify-between md:items-center p-8 px-12 flex-col items-start md:flex-row gap-3">
